+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## ????
-- Custom data parser, better print_r
+- Custom data parser, better_print_r
+- Support for print_r format
 
 ## 0.0.1 - 2023-04-22
 - Basic minimal working version